# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Service Base FastAPI",
    "description": "Version 1 of the FastAPI Application",
    "version": "1.0.0"
  },
  "servers": [{ "url": "/app/v1" }],
  "paths": {
...
    "/organizations/{organization_id}/wallet-transactions": {
      "get": {
        "tags": ["Organization Wallet Transactions"],
        "summary": "Organization Wallet Transactions:List Organization Wallet Transactions",
        "description": "Retrieve a list of organization_wallet transactions with filtering, sorting, and pagination.",
        "operationId": "organization_wallet_transactions_list_organization_wallet_transactions_organizations__organization_id__wallet_transactions_get",
        "security": [{ "OAuth2PasswordBearer": [] }, { "APIKeyCookie": [] }],
        "parameters": [
          {
            "name": "organization_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Organization Id" }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "sorted_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "created_at",
              "title": "Sorted By"
            }
          },
          {
            "name": "sorted_order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "desc",
              "title": "Sorted Order"
            }
          },
          {
            "name": "amount__gte",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "integer" }, { "type": "null" }],
              "title": "Amount  Gte"
            }
          },
          {
            "name": "amount__lte",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "integer" }, { "type": "null" }],
              "title": "Amount  Lte"
            }
          },
          {
            "name": "balance_after_transaction__gte",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "integer" }, { "type": "null" }],
              "title": "Balance After Transaction  Gte"
            }
          },
          {
            "name": "balance_after_transaction__lte",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "integer" }, { "type": "null" }],
              "title": "Balance After Transaction  Lte"
            }
          },
          {
            "name": "wallet_transaction_type__in",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WalletTransactionType"
                  }
                },
                { "type": "null" }
              ],
              "title": "Wallet Transaction Type  In"
            }
          },
          {
            "name": "wallet_transaction_status__in",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WalletTransactionStatus"
                  }
                },
                { "type": "null" }
              ],
              "title": "Wallet Transaction Status  In"
            }
          },
          {
            "name": "created_at__gte",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                { "type": "string", "format": "date-time" },
                { "type": "null" }
              ],
              "title": "Created At  Gte"
            }
          },
          {
            "name": "created_at__lte",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                { "type": "string", "format": "date-time" },
                { "type": "null" }
              ],
              "title": "Created At  Lte"
            }
          },
          {
            "name": "updated_at__gte",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                { "type": "string", "format": "date-time" },
                { "type": "null" }
              ],
              "title": "Updated At  Gte"
            }
          },
          {
            "name": "updated_at__lte",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                { "type": "string", "format": "date-time" },
                { "type": "null" }
              ],
              "title": "Updated At  Lte"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationWalletTransactionListRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation error.",
            "content": {
              "application/json": {
                "examples": {
                  "validation_error": {
                    "summary": "Validation failed on input data.",
                    "value": {
                      "type": "about:blank",
                      "title": "Unprocessable Entity",
                      "status": 422,
                      "instance": "http://127.0.0.1:8000/app/v1/organizations/{organization_id}/wallet-transactions",
                      "errors": [
                        {
                          "status": "422",
                          "code": "validation_error",
                          "title": "Validation Error",
                          "detail": "The input data did not pass validation checks.",
                          "source": { "pointer": "#/body" }
                        }
                      ]
                    }
                  }
                },
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized access.",
            "content": {
              "application/json": {
                "examples": {
                  "unauthorized": {
                    "summary": "Authentication credentials missing or invalid.",
                    "value": {
                      "type": "about:blank",
                      "title": "Unauthorized",
                      "status": 401,
                      "instance": "http://127.0.0.1:8000/app/v1/organizations/{organization_id}/wallet-transactions",
                      "errors": [
                        {
                          "status": "401",
                          "code": "unauthorized",
                          "title": "Unauthorized",
                          "detail": "Authentication credentials were not provided or are invalid."
                        }
                      ]
                    }
                  }
                },
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/organizations/{organization_id}/wallet-transactions/{organization_wallet_transaction_id}": {
      "get": {
        "tags": ["Organization Wallet Transactions"],
        "summary": "Organization Wallet Transactions:Get Organization Wallet Transaction",
        "operationId": "organization_wallet_transactions_get_organization_wallet_transaction_organizations__organization_id__wallet_transactions__organization_wallet_transaction_id__get",
        "security": [{ "OAuth2PasswordBearer": [] }, { "APIKeyCookie": [] }],
        "parameters": [
          {
            "name": "organization_wallet_transaction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Organization Wallet Transaction Id"
            }
          },
          {
            "name": "organization_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Organization Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationWalletTransactionRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation error.",
            "content": {
              "application/json": {
                "examples": {
                  "validation_error": {
                    "summary": "Validation failed on input data.",
                    "value": {
                      "type": "about:blank",
                      "title": "Unprocessable Entity",
                      "status": 422,
                      "instance": "http://127.0.0.1:8000/app/v1/organizations/{organization_id}/wallet-transactions/{organization_wallet_transaction_id}",
                      "errors": [
                        {
                          "status": "422",
                          "code": "validation_error",
                          "title": "Validation Error",
                          "detail": "The input data did not pass validation checks.",
                          "source": { "pointer": "#/body" }
                        }
                      ]
                    }
                  }
                },
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized access.",
            "content": {
              "application/json": {
                "examples": {
                  "unauthorized": {
                    "summary": "Authentication credentials missing or invalid.",
                    "value": {
                      "type": "about:blank",
                      "title": "Unauthorized",
                      "status": 401,
                      "instance": "http://127.0.0.1:8000/app/v1/organizations/{organization_id}/wallet-transactions/{organization_wallet_transaction_id}",
                      "errors": [
                        {
                          "status": "401",
                          "code": "unauthorized",
                          "title": "Unauthorized",
                          "detail": "Authentication credentials were not provided or are invalid."
                        }
                      ]
                    }
                  }
                },
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "properties": {
          "status": { "type": "string", "title": "Status" },
          "code": { "type": "string", "title": "Code" },
          "title": { "type": "string", "title": "Title" },
          "detail": { "type": "string", "title": "Detail" },
          "source": {
            "anyOf": [
              { "$ref": "#/components/schemas/ErrorSource" },
              { "type": "null" }
            ]
          }
        },
        "type": "object",
        "required": ["status", "code", "title", "detail"],
        "title": "ErrorDetail"
      },
      "ErrorResponse": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "default": "about:blank"
          },
          "title": { "type": "string", "title": "Title" },
          "status": { "type": "integer", "title": "Status" },
          "instance": { "type": "string", "title": "Instance" },
          "errors": {
            "items": { "$ref": "#/components/schemas/ErrorDetail" },
            "type": "array",
            "title": "Errors"
          }
        },
        "type": "object",
        "required": ["title", "status", "instance", "errors"],
        "title": "ErrorResponse"
      },
      "ErrorSource": {
        "properties": {
          "pointer": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Pointer"
          }
        },
        "type": "object",
        "title": "ErrorSource"
      },
      "ListResponseMeta": {
        "properties": {
          "limit": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Limit",
            "description": "Maximum number of items to retrieve",
            "default": 100
          },
          "offset": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Offset",
            "description": "Starting position for retrieval",
            "default": 0
          },
          "sorted_by": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Sorted By",
            "description": "Field name to sort by",
            "default": "id"
          },
          "sorted_order": {
            "type": "string",
            "title": "Sorted Order",
            "description": "Sort order: asc or desc",
            "default": "asc"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count",
            "description": "Total number of items matching the search criteria.",
            "example": 100
          }
        },
        "type": "object",
        "required": ["total_count"],
        "title": "ListResponseMeta"
      },
...
      "OrganizationWalletTransactionListRead": {
        "properties": {
          "meta": { "$ref": "#/components/schemas/ListResponseMeta" },
          "data": {
            "items": { "$ref": "#/components/schemas/WalletTransactionRead" },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": ["meta", "data"],
        "title": "OrganizationWalletTransactionListRead"
      },
      "OrganizationWalletTransactionRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "The ID of the object",
            "example": "abcd1234xyzc"
          },
          "amount": {
            "type": "string",
            "title": "Amount",
            "description": "The amount of the transaction in USD",
            "example": "10.00"
          },
          "balance_after_transaction": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Balance After Transaction",
            "description": "The balance after the transaction in USD",
            "example": "100.00"
          },
          "wallet_transaction_type": {
            "$ref": "#/components/schemas/WalletTransactionType",
            "description": "The type of the user_wallet transaction (e.g., 'deposit', 'spend').",
            "example": "deposit"
          },
          "wallet_transaction_status": {
            "$ref": "#/components/schemas/WalletTransactionStatus",
            "description": "The status of the user_wallet transaction (e.g., 'pending', 'completed').",
            "example": "completed"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "The date and time when the transaction was created.",
            "example": "2023-10-01T12:00:00Z"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "The date and time when the transaction was last updated.",
            "example": "2023-10-01T12:00:00Z"
          }
        },
        "type": "object",
        "required": [
          "id",
          "amount",
          "wallet_transaction_type",
          "wallet_transaction_status",
          "created_at",
          "updated_at"
        ],
        "title": "OrganizationWalletTransactionRead"
      },
...
      "WalletTransactionRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "The ID of the object",
            "example": "abcd1234xyzc"
          },
          "amount": {
            "type": "string",
            "title": "Amount",
            "description": "The amount of the transaction in USD",
            "example": "10.00"
          },
          "balance_after_transaction": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Balance After Transaction",
            "description": "The balance after the transaction in USD",
            "example": "100.00"
          },
          "wallet_transaction_type": {
            "$ref": "#/components/schemas/WalletTransactionType",
            "description": "The type of the user_wallet transaction (e.g., 'deposit', 'spend').",
            "example": "deposit"
          },
          "wallet_transaction_status": {
            "$ref": "#/components/schemas/WalletTransactionStatus",
            "description": "The status of the user_wallet transaction (e.g., 'pending', 'completed').",
            "example": "completed"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "The date and time when the transaction was created.",
            "example": "2023-10-01T12:00:00Z"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "The date and time when the transaction was last updated.",
            "example": "2023-10-01T12:00:00Z"
          }
        },
        "type": "object",
        "required": [
          "id",
          "amount",
          "wallet_transaction_type",
          "wallet_transaction_status",
          "created_at",
          "updated_at"
        ],
        "title": "WalletTransactionRead",
        "description": "Schema for reading user_wallet transaction information."
      },
      "WalletTransactionStatus": {
        "type": "string",
        "enum": ["pending", "completed", "failed"],
        "title": "WalletTransactionStatus"
      },
      "WalletTransactionType": {
        "type": "string",
        "enum": ["deposit", "spend"],
        "title": "WalletTransactionType"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": { "password": { "scopes": {}, "tokenUrl": "auth/jwt/login" } }
      },
      "APIKeyCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token"
      }
    }
  }
}